feat: add "idiomatic" rust logging
This introduces the `oldlog` feature to preserve the old logging style, but
default to using the more *idiomatic* logging macros with `pretty_env_logger`
(which should be easy to replace with something else).

This is partially to make the output more readable, and I was informed that the
logging format is kinda an idea that never took off.

Signed-off-by: Christina Sørensen <christina@cafkafk.com>