build(deps): lock all major versions
Currently, several versions are locked on `0`. Because of how semver
works, this means that of these dependencies are allow to freely be
upgraded with breaking changes. Essentially, it's equivalent to setting
`*`, only as long as it doesn't go to a "one dot" release.

Most of the cargo ecosystem is pre one dot, and will be so forever, so
this is extremely brittle.

wharfix v0.1.0 (/home/ces/org/src/dbc/wharfix)
├── actix-rt v2.10.0
├── actix-web v4.9.0
├── async-process v1.8.1
├── async-stream v0.3.5
├── clap v4.4.18
├── dbc-rust-modules v0.1.0 (https://github.com/dbcdk/rust-modules?branch=main#ab38efd1)
├── erased-serde v0.4.5
├── futures v0.3.30
├── get_chunk v1.2.2
├── git2 v0.19.0
├── lazy_static v1.5.0
├── linereader v0.4.0
├── log v0.4.22
├── pretty_env_logger v0.5.0
├── regex v1.10.6
├── serde v1.0.210
├── serde_json v1.0.128
├── sha2 v0.10.8
├── tempdir v0.3.7
├── tempfile v3.12.0
├── tokio v1.40.0
├── uuid v0.8.2
└── walkdir v2.5.0

Signed-off-by: Christina Sørensen <christina@cafkafk.com>